+++ CHANGELOG.md
@@ -8,6 +8,9 @@ version 0.2.0, thus covering only a fraction of the project’s history.
 - `tweaks` is no longer restricted to key aliases.
     - Any named feature can be used in a leaf node.
     - It is no longer necessary to specify a corner or segment.
+- Added a `secondaries` setting for named features that are just points in
+  space, placed in relation to other named features. Used in `tweaks`, these
+  secondaries give greater freedom in shaping the case.
 
 ### Developer
 - Took advantage of new developments in general-purpose libraries:
@@ -16,6 +19,8 @@ version 0.2.0, thus covering only a fraction of the project’s history.
     - `scad-tarmi` lofting replaced `pairwise-hulls` and `triangle-hulls`.
     - `scad-tarmi` flex functions obviated separate functions for object
       placement and reasoning about that placement.
+- Changed the merge order in the `reckon-from-anchor` function to make
+  secondaries useful in tweaks.
 
 ## [Version 0.3.0] - 2019-02-18
 ### Changed